+++ src/main/java/it/manueomm/facefile/utils/Quality.java
@@ -1,4 +1,4 @@
-package it.manueomm.a2pdf.utils;
+package it.manueomm.facefile.utils;
 
 /**
  * Define the image quality choose by the user
